# app/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FEM 巡檢管理系統 API",
    "description": "設施設備巡檢管理系統的 RESTful API 文件",
    "version": "1.0.0",
    "contact": {
      "name": "API 支援",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "API 伺服器"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "輸入您的 JWT Token"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "message": {
            "type": "string",
            "example": "錯誤訊息"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/facilities/tree": {
      "get": {
        "tags": ["設施管理"],
        "summary": "取得設施樹狀結構",
        "description": "取得完整的設施階層樹",
        "parameters": [
          {
            "name": "org_id",
            "in": "query",
            "description": "組織 ID（可選，用於篩選特定組織的設施）",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功取得設施樹",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "facilities": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        },
                        "total_count": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/facilities/list": {
      "get": {
        "tags": ["設施管理"],
        "summary": "取得設施列表",
        "description": "取得扁平化的設施列表",
        "parameters": [
          {
            "name": "org_id",
            "in": "query",
            "description": "組織 ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "parent_id",
            "in": "query",
            "description": "上層設施 ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "facility_type",
            "in": "query",
            "description": "設施類型",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "頁碼",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "每頁筆數",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功取得設施列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "facilities": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        },
                        "pagination": {
                          "type": "object"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/facilities/{facility_id}": {
      "get": {
        "tags": ["設施管理"],
        "summary": "取得設施詳情",
        "parameters": [
          {
            "name": "facility_id",
            "in": "path",
            "required": true,
            "description": "設施 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功取得設施詳情",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "facility": {
                          "type": "object"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "設施不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/facilities/{facility_id}/equipment": {
      "get": {
        "tags": ["設施管理"],
        "summary": "取得設施下的設備",
        "parameters": [
          {
            "name": "facility_id",
            "in": "path",
            "required": true,
            "description": "設施 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功取得設備列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "equipment": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}